Preprocessor divided into smaller classes
